maf_comparator/pyproject.toml
[tool.poetry]
name = "maf_comparator"
version = "0.1.0"
description = "maf-comparator using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = {extras = ["tools"], version = "^0.22.2"}
python-docx = "^1.1.0"

[tool.poetry.scripts]
maf_comparator = "maf_comparator.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"